backend/requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy
web3
requests
langchain_openai
langgraph
cdp_langchain
